srlbackend

+++ app/Http/Controllers/OrdersController.php
@@ -10,6 +10,10 @@ use Purifier;
 
 class OrdersController extends Controller
 {
+  public function __construct()
+{
+    $this->middleware("jwt.auth", ["only" => ["store", "update", "destroy"]]);
+}
   public function index()
   {
     $orders = Order::all();
@@ -34,6 +38,12 @@ class OrdersController extends Controller
     ];
 
     $validator = Validator::make(Purifier::clean($request->all()), $rules)
+
+    $user = Auth::user();
+    if($user->roleID != 1)
+    {
+      return Response::json(['error' => 'Not Authorized']);
+    }
 
     $product = Product::find($request->input("productID");
 
@@ -60,6 +70,23 @@ class OrdersController extends Controller
     return Response::json(['success' => 'successful']);
 
   }
+  public function update($id, Request $request)
+  {
+
+    $order = Order::find($id);
+
+
+    $order->userID = $request->input('userID');
+    $order->productID = $request->input('productID');
+    $order->amount = $request->input('amount');
+    $order->totalPrice = $request->input('totalPrice');
+    $order->comment = $request->input('comment');
+
+    $order->save();
+
+
+    return Response::json(["success" => "Order Updated"]);
+  }
   public function show($id)
   {
     $order = Order::find($id);
@@ -69,12 +96,19 @@ class OrdersController extends Controller
   //Delete a single Articles
   public function destroy($id)
   {
+    $user = Auth::user();
+    if($user->roleID != 1)
+    {
+      return Response::json(['error' => 'Not Authorized']);
+    }
+
     $order = Order::find($id);
 
     $order->delete();
 
     return Response::json(['success' => 'Deleted Article.']);
   }
+
 
 }
